plist: Split out plist_types.h
mainline inclusion
from mainline-v6.8-rc1
category: performance

Trimming down sched.h dependencies: we don't want to include more than
the base types.

Conflicts:
	include/linux/sched.h
(cherry picked from commit 8b7787a543cde905e53eaf29172c9472fe8a6a75)
